Improve tmux functions

+++ dotfiles/.config/fish/functions/tm.fish
@@ -1,10 +1,19 @@
-# Defined in /tmp/fish.l8uIzC/tm.fish @ line 2
+# Defined in /tmp/fish.Ltfyi3/tm.fish @ line 2
 function tm --description 'attach to existing tmux or create new one' --argument session_name
 	test -n "$session_name" || set -l session_name 'def'
-    begin
-        tmux_attach $session_name;
-        or tmux_new $session_name;
-        and kill $fish_pid
-    end;
-    echo "Cannot attach/create tmux session '$session_name'."
+    if type -q tmux
+        if test -z "$TMUX"
+            tmux new-session -A -s "$session_name";
+            and kill $fish_pid;
+        else
+            echo "Already in TMUX session!"
+            return 1
+        end
+    else
+        echo "TMUX not found!"
+        return 1
+    end
+    # Something went wrong
+    echo "Error attaching/creating tmux session '$session_name'."
+    return 1
 end
